+++ seneca.js
@@ -353,10 +353,6 @@ function init(seneca_options, more_options) {
     // that provide functionality, and are thus not really printable
     seneca.log.debug({ kind: 'notice', options: { ...options, internal: null } });
     Print.print_options(seneca, options);
-    // Register default plugins, unless turned off by options.
-    if (options.legacy.transport && options.default_plugins.transport) {
-        seneca.use(require('seneca-transport'));
-    }
     // Register plugins specified in options.
     options.plugins = null == options.plugins ? {} : options.plugins;
     var pluginkeys = Object.keys(options.plugins);
@@ -685,6 +681,8 @@ function make_seneca(initial_opts) {
             process.once(signal, private$.exit_close);
         }
     });
+    // Emit start message (also ensures GateExecutor will trigger 'ready' event)
+    root$.act('sys:seneca,on:point,point:start');
     return root$;
 }
 // Private member variables of Seneca object.