+++ test/logging.test.js
@@ -332,11 +332,7 @@ describe('logging', function () {
         .add('a:1', a1)
         .act('a:1')
         .ready(function () {
-          // hello entry, legacy-transport ready entry
-          // remove legacy-transport entry in 4.x
-
-          expect(log.length).to.equal(2)
-
+          expect(log.length).to.equal(1)
           quiet()
         })
     }
@@ -408,9 +404,7 @@ describe('logging', function () {
         .add('a:1', a1)
         .act('a:1')
         .ready(function () {
-          // hello entry, legacy-transport ready entry
-          // remove legacy-transport entry in 4.x
-          expect(log.length).to.equal(2)
+          expect(log.length).to.equal(1)
 
           json()
         })
@@ -425,7 +419,7 @@ describe('logging', function () {
         .add('a:1', a1)
         .act('a:1')
         .ready(function () {
-          expect(log.length).to.equal(2)
+          expect(log.length).to.equal(1)
 
           flat()
         })
@@ -438,7 +432,7 @@ describe('logging', function () {
         .add('a:1', a1)
         .act('a:1')
         .ready(function () {
-          expect(log.length).to.equal(2)
+          expect(log.length).to.equal(1)
 
           logger_test()
         })
@@ -451,7 +445,7 @@ describe('logging', function () {
         .add('a:1', a1)
         .act('a:1')
         .ready(function () {
-          expect(log.length).to.equal(2)
+          expect(log.length).to.equal(1)
 
           do_test()
         })