+++ test/message.test.js
@@ -158,7 +158,8 @@ describe('message', function () {
 
   it('loop', test_opts, function (fin) {
     var i = 0
-    Seneca({ id$: 'loop', idlen: 4, log: 'silent', limits: { maxparents: 3 } })
+    Seneca({ id$: 'loop', idlen: 4, limits: { maxparents: 3 } })
+      .test()
       .add('a:1', function a1(msg, reply, meta) {
         expect(is_uniq(parents(meta, 1))).true()
 
@@ -172,10 +173,10 @@ describe('message', function () {
         expect(i).equal(4)
         expect(err.code).equal('maxparents')
         expect(err.details.parents).equal([
-          'a:1 a1_8',
-          'a:1 a1_8',
-          'a:1 a1_8',
-          'a:1 a1_8',
+          'a:1 a1_6',
+          'a:1 a1_6',
+          'a:1 a1_6',
+          'a:1 a1_6',
         ])
 
         fin()