+++ test/transport.test.js
@@ -43,6 +43,8 @@ var test_opts = { parallel: false, timeout: 5555 * tmx }
 describe('transport', function () {
   // TODO: test top level qaz:* : def and undef other pats
 
+  // TODO: move to seneca-transport
+  /*
   it('happy-nextgen', test_opts, function (fin) {
     var s0 = Seneca({ id$: 's0', legacy: { transport: false } }).test(fin)
     var c0 = Seneca({
@@ -1099,4 +1101,5 @@ describe('transport', function () {
       })
     }
   )
+  */
 })